Refactoring to be better organized
