+++ src/Core/TAPN/CMakeLists.txt
@@ -4,5 +4,5 @@
 add_library(TAPN InhibitorArc.cpp TimedArcPetriNet.cpp
 TimedPlace.cpp TimeInterval.cpp TransportArc.cpp
 OutputArc.cpp TimedInputArc.cpp TimedTransition.cpp TimeInvariant.cpp
-TAPNModelBuilder.cpp)
+TAPNModelBuilder.cpp StochasticStructure.cpp)
 target_link_libraries(TAPN PlaceVisitor)
